Rename setup services folder
