+++ app/controllers/return-requirements.controller.js
@@ -16,7 +16,7 @@ const ExistingService = require('../services/return-requirements/existing.servic
 const FeatureFlagsConfig = require('../../config/feature-flags.config.js')
 const FrequencyCollectedService = require('../services/return-requirements/frequency-collected.service.js')
 const FrequencyReportedService = require('../services/return-requirements/frequency-reported.service.js')
-const MethodService = require('../services/return-requirements/setup/method.service.js')
+const MethodService = require('../services/return-requirements/method/method.service.js')
 const NoReturnsRequiredService = require('../services/return-requirements/no-returns-required.service.js')
 const NoteService = require('../services/return-requirements/note.service.js')
 const PointsService = require('../services/return-requirements/points.service.js')
@@ -34,7 +34,7 @@ const SubmitCheckService = require('../services/return-requirements/submit-check
 const SubmitExistingService = require('../services/return-requirements/submit-existing.service.js')
 const SubmitFrequencyCollectedService = require('../services/return-requirements/submit-frequency-collected.service.js')
 const SubmitFrequencyReportedService = require('../services/return-requirements/submit-frequency-reported.service.js')
-const SubmitMethodService = require('../services/return-requirements/setup/submit-method.service.js')
+const SubmitMethodService = require('../services/return-requirements/method/submit-method.service.js')
 const SubmitNoReturnsRequiredService = require('../services/return-requirements/submit-no-returns-required.service.js')
 const SubmitNoteService = require('../services/return-requirements/submit-note.service.js')
 const SubmitPointsService = require('../services/return-requirements/submit-points.service.js')
