+++ test/controllers/return-requirements.controller.test.js
@@ -21,7 +21,7 @@ const DeleteNoteService = require('../../app/services/return-requirements/delete
 const ExistingService = require('../../app/services/return-requirements/existing.service.js')
 const FrequencyCollectedService = require('../../app/services/return-requirements/frequency-collected.service.js')
 const FrequencyReportedService = require('../../app/services/return-requirements/frequency-reported.service.js')
-const MethodService = require('../../app/services/return-requirements/setup/method.service.js')
+const MethodService = require('../../app/services/return-requirements/method/method.service.js')
 const NoReturnsRequiredService = require('../../app/services/return-requirements/no-returns-required.service.js')
 const NoteService = require('../../app/services/return-requirements/note.service.js')
 const PointsService = require('../../app/services/return-requirements/points.service.js')
@@ -36,12 +36,12 @@ const SubmitAgreementsExceptions = require('../../app/services/return-requiremen
 const SubmitExistingService = require('../../app/services/return-requirements/submit-existing.service.js')
 const SubmitFrequencyCollectedService = require('../../app/services/return-requirements/submit-frequency-collected.service.js')
 const SubmitFrequencyReportedService = require('../../app/services/return-requirements/submit-frequency-reported.service.js')
+const SubmitMethodService = require('../../app/services/return-requirements/method/submit-method.service.js')
 const SubmitNoReturnsRequiredService = require('../../app/services/return-requirements/submit-no-returns-required.service.js')
 const SubmitPointsService = require('../../app/services/return-requirements/submit-points.service.js')
 const SubmitPurposeService = require('../../app/services/return-requirements/submit-purpose.service.js')
 const SubmitReasonService = require('../../app/services/return-requirements/submit-reason.service.js')
 const SubmitReturnsCycleService = require('../../app/services/return-requirements/submit-returns-cycle.service.js')
-const SubmitMethodService = require('../../app/services/return-requirements/setup/submit-method.service.js')
 const SubmitSiteDescriptionService = require('../../app/services/return-requirements/submit-site-description.service.js')
 const SubmitStartDateService = require('../../app/services/return-requirements/submit-start-date.service.js')
 const ViewService = require('../../app/services/return-requirements/view.service.js')
